moved integration tests to proper session manager spec file
