+++ tests/src/unit/routing-table/routing.spec.ts
@@ -3,18 +3,11 @@
  * @description Unit tests for the Routing Table
  */
 import { expect } from 'chai'
-import { InMemoryKVStore, Configuration, RoutingTable, PocketRpcProvider, Pocket, Account, PocketAAT,typeGuard, V1RPCRoutes, UnlockedAccount } from '../../../../src'
+import { InMemoryKVStore, Configuration, RoutingTable } from '../../../../src'
 // Constants
 // For Testing we are using dummy data, none of the following information is real.
 const dispatcher = new URL("http://localhost:8081")
 const store = new InMemoryKVStore()
-const configuration = new Configuration(30, 1000, undefined, 40000, true, undefined, undefined, undefined, undefined, false)
-//
-const clientPrivateKey = ""
-const walletAppPublicKey = ""
-const walletPrivateKey = ""
-const clientPassphrase = "123"
-const blockchain = "0002"
 
 describe('Routing Table tests',() => {
     it('should initialize a routing table', () => {
@@ -103,75 +96,6 @@ describe('Routing Table tests',() => {
         expect(readDispatchers[0]).to.be.an.instanceof(URL)
         expect(readDispatchers[1]).to.be.an.instanceof(URL)
         expect(readDispatchers[2]).to.be.an.instanceof(URL)
-    }).timeout(0)
-
-    it("should be able to increase the routing table dispatchers count by performing relay requests", async () => {                
-        // Instantiate Pocket
-        const pocket = new Pocket([dispatcher], undefined, configuration)
-
-        // Import Client Account
-        const clientAccountOrError = await pocket.keybase.importAccount(Buffer.from(clientPrivateKey, "hex"), clientPassphrase)
-        const clientAccount = clientAccountOrError as Account
-        
-        // Unlock client account
-        await pocket.keybase.unlockAccount(clientAccount.addressHex, clientPassphrase, 0)
-    
-        // Create AAT for imported account
-        const aat = await PocketAAT.from("0.0.1", clientAccount.publicKey.toString("hex"), walletAppPublicKey, walletPrivateKey)
-
-        const provider = new PocketRpcProvider(pocket, aat, blockchain, false)
-
-        const firstDispatchersCount = pocket.getDispatchersCount()
-
-        await pocket.rpc(provider)!.query.getHeight()
-
-        const secondDispatchersCount = pocket.getDispatchersCount()
-
-        await pocket.rpc(provider)!.query.getHeight()
-    
-        const thirdDispatchersCount = pocket.getDispatchersCount()
-
-        expect(firstDispatchersCount).to.be.below(secondDispatchersCount).to.be.below(thirdDispatchersCount)
-    }).timeout(0)
-
-    it("should be able to increase the routing table dispatchers count by performing relay requests for two different client accounts", async () => {                
-        // Instantiate Pocket
-        const pocket = new Pocket([dispatcher], undefined, configuration)
-
-        // Import Client Account
-        const clientAccountOrError = await pocket.keybase.importAccount(Buffer.from(clientPrivateKey, "hex"), clientPassphrase)
-        const clientAccount = clientAccountOrError as Account
-        // Unlock client account
-        await pocket.keybase.unlockAccount(clientAccount.addressHex, clientPassphrase, 0)
-        // Create AAT for imported account
-        const aat = await PocketAAT.from("0.0.1", clientAccount.publicKey.toString("hex"), walletAppPublicKey, walletPrivateKey)
-        // Create the PocketRpcProvider for the first client account
-        const provider = new PocketRpcProvider(pocket, aat, blockchain, false)
-
-        // Create a second client account --
-        const clientAccountOrError2 = await pocket.keybase.createAccount(clientPassphrase)
-        const clientAccount2 = clientAccountOrError2 as Account
-        // Unlock client account
-        await pocket.keybase.unlockAccount(clientAccount2.addressHex, clientPassphrase, 0)
-        // Create AAT for imported account
-        const aat2 = await PocketAAT.from("0.0.1", clientAccount.publicKey.toString("hex"), walletAppPublicKey, walletPrivateKey)
-        // Create the PocketRpcProvider for the second client account
-        const provider2 = new PocketRpcProvider(pocket, aat2, blockchain, false)
-
-        // Get the first dispatchers count
-        const firstDispatchersCount = pocket.getDispatchersCount()
-        // Perform a relay request using the first provider that contains the first account aat information
-        await pocket.rpc(provider)!.query.getHeight()
-
-        // Get the second dispatchers count
-        const secondDispatchersCount = pocket.getDispatchersCount()
-        // Perform a relay request using the second provider that contains the second account aat information
-        await pocket.rpc(provider2)!.query.getHeight()
-    
-        // Get the third dispatchers count
-        const thirdDispatchersCount = pocket.getDispatchersCount()
-
-        expect(firstDispatchersCount).to.be.below(secondDispatchersCount).to.be.below(thirdDispatchersCount)
     }).timeout(0)
 
 }).timeout(0)
